Refactored some tests.
